refactor(scripts): give the app-floor parser one home
The declared macOS floor is read by two guards — the Package.swift check here
and before-pack's pack-time payload check — and a second copy of the parser is
exactly how one ends up hardened and the other not. That already happened: the
Package.swift parser was scoped to its declaration block after review while the
function written directly beside it still took the first match in the file.

Moves the string-aware comment strip and the brace matcher into
scripts/macos-floor.mjs so there is one implementation to harden. No behaviour
change; the decoy cases move with it.

Refs #515
